requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.1
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
blinker==1.5
cachetools==5.3.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.0
cryptography==41.0.2
Django==4.2.3
dlib==19.24.2
dnspython==2.4.1
dominate==2.8.0
dropbox==11.36.2
email-validator==2.0.0.post2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-User==1.0.2.2
Flask-WTF==1.1.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gphotospy==0.2.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
image==1.5.33
importlib-metadata==6.8.0
imutils==0.5.4
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
keyring==24.2.0
keyrings.alt==5.0.0
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.0.0
multidict==6.0.4
networkx==3.1
numpy==1.25.1
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.8.0.74
pandas==2.0.3
passlib==1.7.4
Pillow==10.0.0
pillow-heif==0.13.0
ply==3.11
priority==2.0.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pyicloud==1.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
quart==0.18.4
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
stone==3.3.1
toml==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==1.26.16
URLObject==2.4.3
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.8
Werkzeug==2.3.6
wsproto==1.2.0
WTForm==1.0
wtform-address==0.1.1
WTForms==3.0.1
yarl==1.9.2
zipp==3.16.2
